refactor(reviews): zera react-doctor em MyVerdictsView (#232)
Onda 4 (piloto) da campanha "Zerar react-doctor" (#239). Elimina os 9
diagnosticos do arquivo (2 errors + 7 warnings) refatorando o debito
real, sem supressoes. Valida em uso real os hooks compartilhados da
Onda 3 (#231).

- Deriva docIndex de selectedDocId via useMemo (molde AutoReviewPage),
  removendo o effect que resetava o indice ao mudar filtro
  (no-adjust-state-on-prop-change x2, no-derived-state-effect,
  no-chain-state-updates, no-effect-chain).
- Consome useDocumentText no lugar do lazy-load manual
  (no-cascading-set-state, exhaustive-deps).
- Extrai VerdictsHeader e VerdictsList para arquivos-irmaos,
  encolhendo MyVerdictsViewInner de 306 p/ ~150 linhas e os useState
  de 8 p/ 4 (no-giant-component, prefer-useReducer).
- selectRespondent passa a usar useUrlState.

O effect de scroll e o state de comentario migram para VerdictsList,
co-localizados com o DOM/UI que controlam. Comportamento preservado;
mudanca intencional: ao filtrar, o doc selecionado e mantido se
sobrevive ao filtro (antes resetava sempre para o 1o).

react-doctor: 0 diagnosticos nos 3 arquivos; --diff vs main limpo.
tsc e eslint limpos.
